Move logic into provider, improve regex for handle matching, cleanup

+++ providers/default.rb
@@ -191,11 +191,20 @@ def chef_backup_dir
   "#{chef_install_dir}.upgrade"
 end
 
+def chef_upgrade_log
+  "#{chef_install_dir}_upgrade.log"
+end
+
 # cleanup cruft from *prior* runs
 def cleanup
   if ::File.exist?(chef_backup_dir)
     converge_by("remove #{chef_backup_dir} from previous chef-client run") do
       FileUtils.rm_rf chef_backup_dir
+    end
+  end
+  if ::File.exist?(chef_upgrade_log)
+    converge_by("remove #{chef_upgrade_log} from previous chef-client run") do
+      FileUtils.rm_rf chef_upgrade_log
     end
   end
   # When running under init this cron job is created after an update
@@ -291,7 +300,7 @@ def open_handle_functions
     )
     $handleOutput = &#{Chef::Config[:file_cache_path]}/handle.exe -accepteula -nobanner -a -u $Search
     $handleOutput | foreach {
-      if ($_ -match '^(?<program>\\S*)\\s*pid: (?<pid>\\d*)\\s*type: (?<type>\\S*)\\s*(?<user>\\S*)\\s*(?<handle>\\d*):\\s*(?<file>(\\\\\\\\)|([a-z]:).*)') {
+      if ($_ -match '^(?<program>\\S*)\\s*pid: (?<pid>\\d*)\\s*type: (?<type>\\S*)\\s*(?<user>\\S*)\\s*(?<handle>\\S*):\\s*(?<file>(\\\\\\\\)|([a-z]:).*)') {
         $matches | select @{n="User";e={$_.user}},@{n="Path";e={$_.file}},@{n="Handle";e={$_.handle}},@{n="Type";e={$_.type}},@{n="HandlePid";e={$_.pid}},@{n="Program";e={$_.program}}
       }
     }
@@ -335,6 +344,11 @@ def execute_install_script(install_script)
                     ''
                   end
 
+    remote_file "#{Chef::Config[:file_cache_path]}/handle.zip" do
+      source new_resource.handle_zip_download_url
+      not_if { ::File.file?("#{Chef::Config[:file_cache_path]}/handle.exe") }
+    end.run_action(:create)
+
     powershell_script 'name' do
       code <<-EOH
         $command = {
@@ -353,12 +367,18 @@ def execute_install_script(install_script)
 
           #{open_handle_functions}
 
-          Destroy-OpenChefHandles
+          if (Test-Path "#{Chef::Config[:file_cache_path]}/handle.exe") {
+            Destroy-OpenChefHandles
+          }
+
           Get-Service EventLog | Restart-Service -Force
 
           Remove-Item "#{chef_install_dir}" -Recurse -Force
 
-          if (test-path "#{chef_install_dir}") { exit 3 }
+          if (test-path "#{chef_install_dir}") {
+            Write-Output "#{chef_install_dir} still exists, upgrade will be aborted. Exiting (3)..."
+            exit 3
+          }
 
           #{uninstall_first}
           #{install_script}
